git-svn-id: svn://svn.code.sf.net/p/codeblocks/code/branches/scintilla_3_5_x@10593 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/wxSmithPlot/wxsVector.h
@@ -23,7 +23,7 @@
 #include    <iostream>
 
 #include    <wx/stattext.h>
-#include    "../wxswidget.h"
+#include    <wxwidgets/wxswidget.h>
 #include    <mathplot.h>
 
 
@@ -53,8 +53,8 @@ class wxsVector: public wxsWidget
 
         wxArrayString       mXs;            // internal use; X data as a string
         wxArrayString       mYs;            // Y data as a string
-        std::vector<float>  mXf;            // X data as a float value
-        std::vector<float>  mYf;            // Y data as a float
+        std::vector<double>  mXf;            // X data as a double value
+        std::vector<double>  mYf;            // Y data as a double
 
 };
